quete input en cours
