+++ src/app/app.routes.ts
@@ -1,13 +1,13 @@
 import { Routes } from '@angular/router';
 import { HomePageComponent } from './pages/home-page/home-page.component';
-import { ArticlePageComponent } from './pages/article-page/article-page.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
 import { ContactFormComponent } from './pages/contact-form/contact-form.component';
 import { SignupFormComponent } from './pages/signup-form/signup-form.component';
+import { ArticleComponent } from './article/article.component';
 
 export const routes: Routes = [
-    { path: 'article/:id', component: ArticlePageComponent },  
     { path: '', component: HomePageComponent},
+    { path: 'article/:id', component: ArticleComponent },  
     { path: 'contact', component: ContactFormComponent },
     { path: 'registration', component: SignupFormComponent },
