--- src/app/article-list/article-list.component.scss
@@ -1,25 +1,0 @@
-.example-card {
-    max-width: 400px;
-    border: 1px solid #e0e0e0;
-    border-radius: 8px;
-  }
-  
-
-/* Style spécial pour articles populaires */
-.popular-article {
-  font-size: 1.5em;
-  text-shadow: 2px 2px 5px gray;
-  font-weight: bold;
-}
-
-/* *** */
-.lign {
-  display: flex;
-  flex-direction: row;
-  text-align: center;
-  justify-content: center;
-}
-
-.icon {
-  padding: 0 5px;
-}
