+++ src/app/models/article.models.ts
@@ -1,0 +1,10 @@
+export interface Article {
+    title: string;
+    content: string;
+    image: string;
+    author: string;
+    isPublished: boolean;
+    likes: number;
+    comment?: string;
+    id: number;
+  }
