--- src/app/pages/article-page/article-page.component.html
@@ -1,1 +1,0 @@
-<p>Fiche produit de l'article {{ articleId }}</p>
