+++ src/app/pages/home-page/home-page.component.html
@@ -1,3 +1,3 @@
 <div class="contentWrapper">
-  <app-article></app-article>
+  <app-article-list></app-article-list>
 </div>
