+++ src/app/pages/home-page/home-page.component.scss
@@ -1,8 +1,8 @@
 /* Wrapper */
 .contentWrapper {
     display: flex;
+    flex-direction: row;
     flex-wrap: wrap;
-    gap: 16px;
     justify-content: center;
     text-align: center;
 }
